+++ CHANGELOG.md
@@ -7,8 +7,11 @@ All notable changes to this project will be documented in this file.
 - Add filter info to graph label(Confirm necessity because it requires graph redesign)
 - Chart for the population distribution
 
-### Upcoming Next Release
-- Influence Reach page with last post and only opinion authors
+## [1.1] 2016-07-28
+### Added New DB.sql Import needed
+- Loading ready for over-time import
+- Loading set to month to month to help with influence calc
+Influence Reach page with last post and only opinion authors
 
 ## [1.0.6] 2016-07-27 Minor Fix on 2016-07-28
 ### Added/Fixed